refactor(all): restructure directory structure further
