+++ src/client/LodestoneClient.ts
@@ -26,14 +26,14 @@
 import Axios, { AxiosError, AxiosInstance } from 'axios'
 import Cheerio, { CheerioAPI } from 'cheerio'
 import pLimit, { Limit } from 'p-limit'
-import Character from '../character/Character'
-import Worlds from '../world/Worlds'
+import Character from '../entity/character/Character'
+import Worlds from '../entity/world/Worlds'
 import CharacterNotFoundError from '../errors/CharacterNotFoundError'
 import { ICharacterFetchError } from './interface/ICharacterFetchError'
 import CharacterFetchError from '../errors/CharacterFetchError'
 import CharacterFetchTimeoutError from '../errors/CharacterFetchTimeoutError'
 import ICharacterSetFetchResult from './interface/ICharacterSetFetchResult'
-import Creature from '../creature/Creature'
+import Creature from '../entity/creature/Creature'
 import IClientProps from './interface/IClientProps'
 import Language from '../locale/Language'
 import LocalizedClientFactory from '../locale/LocalizedClientFactory'
