+++ src/entity/character/Character.ts
@@ -27,14 +27,14 @@ import { CheerioAPI } from 'cheerio'
 import CharacterDomConfig from './config/CharacterDomConfig'
 import ICharacter from './interface/ICharacter'
 import IGearSet from './attribute/gear/IGearSet'
-import IAttributeMapping from '../interface/IAttributeMapping'
+import IAttributeMapping from '../../parser/interface/IAttributeMapping'
 import ClassAbbreviation from './attribute/class/category/ClassAbbreviation'
 import GearCategory from './attribute/gear/GearCategory'
 import Level from './attribute/class/Level'
-import IItem from '../interface/IItem'
+import IItem from '../item/interface/IItem'
 import IPlayerGroup from './attribute/group/IPlayerGroup'
-import UnparseableGroupIdError from '../errors/UnparseableGroupIdError'
-import Language from '../locale/Language'
+import UnparseableGroupIdError from '../../errors/UnparseableGroupIdError'
+import Language from '../../locale/Language'
 import ClassLevels from './attribute/class/ClassLevels'
 
 export default class Character implements ICharacter {
