+++ src/entity/character/__tests__/resources/18001255.html
