+++ src/entity/character/__tests__/resources/27218992.html
