+++ src/entity/character/__tests__/resources/38531003.html
