+++ src/entity/character/__tests__/resources/8283.html
