+++ src/entity/character/attribute/class/ClassInfoMap.ts
