+++ src/entity/character/attribute/class/ClassLevels.ts
