+++ src/entity/character/attribute/class/category/ClassAbbreviation.ts
