+++ src/entity/character/attribute/class/category/ClassCategory.ts
