+++ src/entity/character/attribute/class/category/Role.ts
