+++ src/entity/character/attribute/class/interface/ILevel.ts
