+++ src/entity/character/attribute/gear/GearCategory.ts
