+++ src/entity/character/attribute/gear/IGearSet.ts
