+++ src/entity/character/attribute/group/IPlayerGroup.ts
