+++ src/entity/character/config/CharacterDomConfig.ts
