+++ src/entity/character/config/ClassConfig.ts
