+++ src/entity/creature/Creature.ts
@@ -23,11 +23,31 @@
  *
  */
 
-import GearCategory from './GearCategory'
-import IItem from '../../../interface/IItem'
+import { CheerioAPI } from 'cheerio'
+import IItem from '../item/interface/IItem'
+import CreatureCategory from './type/CreatureCategory'
 
-export default interface IGearPiece extends IItem {
-  name: string
-  category: GearCategory
-  iLvl: number
+export default class Creature {
+  constructor(public toolTipId: string, public item: IItem, public type?: CreatureCategory, public name?: string) {}
+
+  public asMapping(): { toolTipId: string; itemId: string } {
+    return {
+      toolTipId: this.toolTipId,
+      itemId: this.item.id,
+    }
+  }
+
+  static fromToolTip(toolTipId: string, data: string, cheerio: CheerioAPI, itemIdOnly?: boolean): Creature {
+    const $ = cheerio.load(data)
+    const item = $('a')[0]
+    return new Creature(
+      toolTipId,
+      {
+        name: !itemIdOnly ? item.attribs['data-tooltip'] : undefined,
+        id: item.attribs.href.replace('/lodestone/playguide/db/item/', '').replace('/', ''),
+      },
+      $('header')[0].attribs.class.replace('__header', '').toUpperCase() as CreatureCategory,
+      !itemIdOnly ? $('h4').text() : undefined
+    )
+  }
 }
