+++ src/entity/creature/__tests__/Creature.test.ts
