+++ src/entity/creature/type/CreatureCategory.ts
