+++ src/entity/item/interface/IItem.ts
