+++ src/entity/world/__tests__/resources/worldstatus.html
