+++ src/entity/world/attribute/Region.ts
