+++ src/entity/world/attribute/WorldCategory.ts
