+++ src/entity/world/attribute/WorldStatus.ts
