+++ src/entity/world/interface/IDataCenter.ts
