+++ src/entity/world/interface/IWorld.ts
