+++ src/parser/interface/IAttributeMapping.ts
